+++ examples/resources/elasticstack_elasticsearch_ingest_pipeline/import.sh
@@ -1,0 +1,1 @@
+terraform import elasticstack_elasticsearch_ingest_pipeline.my_ingest_pipeline <cluster_uuid>/<ingest pipeline name>